@@ -10,40 +10,40 @@
   },
   "dependencies": {
     "@babel/helper-regex": "^7.10.5",
-    "@react-google-maps/api": "^2.12.0",
-    "@reduxjs/toolkit": "^1.8.0",
-    "axios": "^0.24.0",
-    "bootstrap": "^5.1.3",
-    "react": "^17.0.2",
+    "@react-google-maps/api": "^2.18.1",
+    "@reduxjs/toolkit": "^1.9.3",
+    "axios": "^1.3.4",
+    "bootstrap": "^5.2.3",
+    "react": "^18.2.0",
     "react-beautiful-dnd": "^13.1.1",
-    "react-bootstrap": "^2.0.3",
-    "react-confirm-alert": "^2.7.0",
+    "react-bootstrap": "^2.7.2",
+    "react-confirm-alert": "^3.0.6",
     "react-ctrl-f": "^0.0.4",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-geocode": "^0.2.3",
-    "react-google-recaptcha-v3": "^1.9.7",
-    "react-icons": "^4.3.1",
-    "react-redux": "^7.2.6",
-    "react-router": "^6.0.2",
-    "react-router-dom": "^6.0.2",
+    "react-google-recaptcha-v3": "^1.10.1",
+    "react-icons": "^4.8.0",
+    "react-redux": "^8.0.5",
+    "react-router": "^6.8.2",
+    "react-router-dom": "^6.8.2",
     "react-scripts": "^5.0.1",
-    "react-share": "^4.4.0",
-    "socket.io-client": "^4.3.2"
+    "react-share": "^4.4.1",
+    "socket.io-client": "^4.6.1"
   },
   "devDependencies": {
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.11.7",
-    "@types/react": "^18.0.14",
-    "@types/react-beautiful-dnd": "^13.1.2",
-    "@types/react-dom": "^18.0.5",
+    "@types/jest": "^29.4.0",
+    "@types/node": "^18.14.6",
+    "@types/react": "^18.0.28",
+    "@types/react-beautiful-dnd": "^13.1.3",
+    "@types/react-dom": "^18.0.11",
     "@types/react-geocode": "^0.2.1",
     "@types/react-icons": "^3.0.0",
-    "@types/react-redux": "^7.1.24",
-    "@types/react-router": "^5.1.17",
-    "@types/react-router-dom": "^5.3.2",
+    "@types/react-redux": "^7.1.25",
+    "@types/react-router": "^5.1.20",
+    "@types/react-router-dom": "^5.3.3",
     "@types/socket.io-client": "^3.0.0",
-    "google-auth-library": "^8.1.1",
-    "typescript": "4.3.5"
+    "google-auth-library": "^8.7.0",
+    "typescript": "4.9.5"
   },
   "eslintConfig": {
     "extends": "react-app"
